web-monitor/backend/requirements.txt
fastapi
uvicorn
sqlmodel
paho-mqtt
python-dotenv
pydantic
requests
httpx
scikit-learn
pandas
joblib
